requirements.txt
Flask==2.1.3
waitress==2.1.2
flask-cors
python-dotenv
werkzeug==0.16.1